Modify awesome books to use es6 Use modules folder to divide code into junks Use export and import Refactor functions and methods to arrow functions Modify Date function to use Luxon library
