+++ index.html
@@ -18,10 +18,11 @@
         </ul>
       </nav>
     </header>
+    <span class="date"></span>
     <main class="main-section"></main>
     <footer>
       <p>Copyright © Yasin</p>
     </footer>
-    <script type="text/javascript" src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
